requirements.txt
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
future==1.0.0
imageio==2.34.2
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pillow==10.4.0
scikit-image==0.24.0
scipy==1.13.1
sympy==1.13.1
tifffile==2024.7.24
torch==1.10.1+cpu
torchaudio==0.10.1+cpu
torchvision==0.11.2+cpu
tqdm==4.66.4
typing_extensions==4.12.2
